HTML {shiny}	R Documentation

Mark Characters as HTML

Description

Marks the given text as HTML, which means the tag functions will know not to perform HTML escaping on it.

Usage

HTML(text, ...)

Arguments

text
The text value to mark with HTML

...
Any additional values to be converted to character and concatenated together

Value

The same value, but marked as HTML.

Examples

el <- div(HTML("I like <u>turtles</u>"))
cat(as.character(el))

[Package shiny version 1.2.0 Index]