applyInputHandlers {shiny}	R Documentation

Apply input handlers to raw input values

Description

The purpose of this function is to make it possible for external packages to test Shiny inputs. It takes a named list of raw input values, applies input handlers to those values, and then returns a named list of the processed values.

Usage

applyInputHandlers(inputs, shinysession = getDefaultReactiveDomain())

Arguments

inputs
A named list of input values.

shinysession
A Shiny session object.

Details

The raw input values should be in a named list. Some values may have names like "x:shiny.date". This function would apply the "shiny.date" input handler to the value, and then rename the result to "x", in the output.

See Also

registerInputHandler

[Package shiny version 1.9.1 Index]