dateRangeInput {shiny}	R Documentation

Create date range input

Description

Creates a pair of text inputs which, when clicked on, bring up calendars that the user can click on to select dates.

Usage

dateRangeInput(inputId, label, start = NULL, end = NULL, min = NULL,
  max = NULL, format = "yyyy-mm-dd", startview = "month",
  weekstart = 0, language = "en", separator = " to ", width = NULL,
  autoclose = TRUE)

Arguments

inputId
The input slot that will be used to access the value.

label
Display label for the control, or NULL for no label.

start
The initial start date. Either a Date object, or a string in yyyy-mm-dd format. If NULL (the default), will use the current date in the client's time zone.

end
The initial end date. Either a Date object, or a string in yyyy-mm-dd format. If NULL (the default), will use the current date in the client's time zone.

min
The minimum allowed date. Either a Date object, or a string in yyyy-mm-dd format.

max
The maximum allowed date. Either a Date object, or a string in yyyy-mm-dd format.

format
The format of the date to display in the browser. Defaults to "yyyy-mm-dd".

startview
The date range shown when the input object is first clicked. Can be "month" (the default), "year", or "decade".

weekstart
Which day is the start of the week. Should be an integer from 0 (Sunday) to 6 (Saturday).

language
The language used for month and day names. Default is "en". Other valid values include "ar", "az", "bg", "bs", "ca", "cs", "cy", "da", "de", "el", "en-AU", "en-GB", "eo", "es", "et", "eu", "fa", "fi", "fo", "fr-CH", "fr", "gl", "he", "hr", "hu", "hy", "id", "is", "it-CH", "it", "ja", "ka", "kh", "kk", "ko", "kr", "lt", "lv", "me", "mk", "mn", "ms", "nb", "nl-BE", "nl", "no", "pl", "pt-BR", "pt", "ro", "rs-latin", "rs", "ru", "sk", "sl", "sq", "sr-latin", "sr", "sv", "sw", "th", "tr", "uk", "vi", "zh-CN", and "zh-TW".

separator
String to display between the start and end input boxes.

width
The width of the input, e.g. '400px', or '100%'; see validateCssUnit.

autoclose
Whether or not to close the datepicker immediately when a date is selected.

Details

The date format string specifies how the date will be displayed in the browser. It allows the following values:

yy Year without century (12)

yyyy Year with century (2012)

mm Month number, with leading zero (01-12)

m Month number, without leading zero (1-12)

M Abbreviated month name

MM Full month name

dd Day of month with leading zero

d Day of month without leading zero

D Abbreviated weekday name

DD Full weekday name

See Also

dateInput, updateDateRangeInput

Other input elements: actionButton, checkboxGroupInput, checkboxInput, dateInput, fileInput, numericInput, passwordInput, radioButtons, selectInput, sliderInput, submitButton, textAreaInput, textInput, varSelectInput

Examples

## Only run examples in interactive R sessions
if (interactive()) {

ui <- fluidPage(
  dateRangeInput("daterange1", "Date range:",
                 start = "2001-01-01",
                 end   = "2010-12-31"),

  # Default start and end is the current date in the client's time zone
  dateRangeInput("daterange2", "Date range:"),

  # start and end are always specified in yyyy-mm-dd, even if the display
  # format is different
  dateRangeInput("daterange3", "Date range:",
                 start  = "2001-01-01",
                 end    = "2010-12-31",
                 min    = "2001-01-01",
                 max    = "2012-12-21",
                 format = "mm/dd/yy",
                 separator = " - "),

  # Pass in Date objects
  dateRangeInput("daterange4", "Date range:",
                 start = Sys.Date()-10,
                 end = Sys.Date()+10),

  # Use different language and different first day of week
  dateRangeInput("daterange5", "Date range:",
                 language = "de",
                 weekstart = 1),

  # Start with decade view instead of default month view
  dateRangeInput("daterange6", "Date range:",
                 startview = "decade")
)

shinyApp(ui, server = function(input, output) { })
}

[Package shiny version 1.2.0 Index]