dblclickOpts {shiny}	R Documentation

Create an object representing double-click options

Description

This generates an object representing dobule-click options, to be passed as the dblclick argument of imageOutput or plotOutput.

Usage

dblclickOpts(id = NULL, clip = TRUE, delay = 400)

Arguments

id
Input value name. For example, if the value is "plot_dblclick", then the click coordinates will be available as input$plot_dblclick.

clip
Should the click area be clipped to the plotting area? If FALSE, then the server will receive double-click events even when the mouse is outside the plotting area, as long as it is still inside the image.

delay
Maximum delay (in ms) between a pair clicks for them to be counted as a double-click.

[Package shiny version 1.2.0 Index]