exprToFunction {shiny}	R Documentation

Convert an expression to a function

Description

Please use installExprFunction() for a better debugging experience (Shiny 0.8.0). If the expr and quoted parameters are not needed, please see quoToFunction() (Shiny 1.6.0).

Usage

exprToFunction(expr, env = parent.frame(), quoted = FALSE)

Arguments

expr
A quoted or unquoted expression, or a quosure.

env
The desired environment for the function. Defaults to the calling environment two steps back.

quoted
Is the expression quoted?

Details

Similar to installExprFunction() but doesn't register debug hooks.

See Also

installExprFunction() for the modern approach to converting an expression to a function

[Package shiny version 1.9.1 Index]