getCurrentOutputInfo {shiny}	R Documentation

Get information about the output that is currently being executed.

Description

Get information about the output that is currently being executed.

Usage

getCurrentOutputInfo(session = getDefaultReactiveDomain())

Arguments

session
The current Shiny session.

[Package shiny version 1.2.0 Index]