hoverOpts {shiny}	R Documentation

Create an object representing hover options

Description

This generates an object representing hovering options, to be passed as the hover argument of imageOutput or plotOutput.

Usage

hoverOpts(id = NULL, delay = 300, delayType = c("debounce",
  "throttle"), clip = TRUE, nullOutside = TRUE)

Arguments

id
Input value name. For example, if the value is "plot_hover", then the hover coordinates will be available as input$plot_hover.

delay
How long to delay (in milliseconds) when debouncing or throttling, before sending the mouse location to the server.

delayType
The type of algorithm for limiting the number of hover events. Use "throttle" to limit the number of hover events to one every delay milliseconds. Use "debounce" to suspend events while the cursor is moving, and wait until the cursor has been at rest for delay milliseconds before sending an event.

clip
Should the hover area be clipped to the plotting area? If FALSE, then the server will receive hover events even when the mouse is outside the plotting area, as long as it is still inside the image.

nullOutside
If TRUE (the default), the value will be set to NULL when the mouse exits the plotting area. If FALSE, the value will stop changing when the cursor exits the plotting area.

[Package shiny version 1.2.0 Index]