mainPanel {shiny}	R Documentation

Create a main panel

Description

Create a main panel containing output elements that can in turn be passed to sidebarLayout.

Usage

mainPanel(..., width = 8)

Arguments

...
Output elements to include in the main panel

width
The width of the main panel. For fluid layouts this is out of 12 total units; for fixed layouts it is out of whatever the width of the main panel's parent column is.

Value

A main panel that can be passed to sidebarLayout.

Examples

# Show the caption and plot of the requested variable against mpg
mainPanel(
   h3(textOutput("caption")),
   plotOutput("mpgPlot")
)

[Package shiny version 1.2.0 Index]