pageWithSidebar {shiny}	R Documentation

Create a page with a sidebar

Description

Create a Shiny UI that contains a header with the application title, a sidebar for input controls, and a main area for output.

Usage

pageWithSidebar(headerPanel, sidebarPanel, mainPanel)

Arguments

headerPanel
The headerPanel with the application title

sidebarPanel
The sidebarPanel containing input controls

mainPanel
The mainPanel containing outputs

Value

A UI defintion that can be passed to the shinyUI function

Note

This function is deprecated. You should use fluidPage along with sidebarLayout to implement a page with a sidebar.

Examples

# Define UI
pageWithSidebar(

  # Application title
  headerPanel("Hello Shiny!"),

  # Sidebar with a slider input
  sidebarPanel(
    sliderInput("obs",
                "Number of observations:",
                min = 0,
                max = 1000,
                value = 500)
  ),

  # Show a plot of the generated distribution
  mainPanel(
    plotOutput("distPlot")
  )
)

[Package shiny version 1.2.0 Index]