reactiveConsole {shiny}	R Documentation

Activate reactivity in the console

Description

This is an experimental feature that allows you to enable reactivity at the console, for the purposes of experimentation and learning.

Usage

reactiveConsole(enabled)

Arguments

enabled
Turn console reactivity on or off?

Examples

reactiveConsole(TRUE)
x <- reactiveVal(10)
y <- observe({
  message("The value of x is ", x())
})
x(20)
x(30)
reactiveConsole(FALSE)

[Package shiny version 1.9.1 Index]