removeInputHandler {shiny}	R Documentation

Deregister an Input Handler

Description

Removes an Input Handler. Rather than using the previously specified handler for data of this type, the default jsonlite serialization will be used.

Usage

removeInputHandler(type)

Arguments

type
The type for which handlers should be removed.

Value

The handler previously associated with this type, if one existed. Otherwise, NULL.

See Also

registerInputHandler()

[Package shiny version 1.9.1 Index]