renderTable {shiny}	R Documentation

Table Output

Description

Creates a reactive table that is suitable for assigning to an output slot.

Usage

renderTable(expr, striped = FALSE, hover = FALSE, bordered = FALSE,
  spacing = c("s", "xs", "m", "l"), width = "auto", align = NULL,
  rownames = FALSE, colnames = TRUE, digits = NULL, na = "NA", ...,
  env = parent.frame(), quoted = FALSE, outputArgs = list())

Arguments

expr
An expression that returns an R object that can be used with xtable.

striped, hover, bordered
Logicals: if TRUE, apply the corresponding Bootstrap table format to the output table.

spacing
The spacing between the rows of the table (xs stands for "extra small", s for "small", m for "medium" and l for "large").

width
Table width. Must be a valid CSS unit (like "100 "auto") or a number, which will be coerced to a string and have "px" appended.

align
A string that specifies the column alignment. If equal to 'l', 'c' or 'r', then all columns will be, respectively, left-, center- or right-aligned. Otherwise, align must have the same number of characters as the resulting table (if rownames = TRUE, this will be equal to ncol()+1), with the i-th character specifying the alignment for the i-th column (besides 'l', 'c' and 'r', '?' is also permitted - '?' is a placeholder for that particular column, indicating that it should keep its default alignment). If NULL, then all numeric/integer columns (including the row names, if they are numbers) will be right-aligned and everything else will be left-aligned (align = '?' produces the same result).

rownames, colnames
Logicals: include rownames? include colnames (column headers)?

digits
An integer specifying the number of decimal places for the numeric columns (this will not apply to columns with an integer class). If digits is set to a negative value, then the numeric columns will be displayed in scientific format with a precision of abs(digits) digits.

na
The string to use in the table cells whose values are missing (i.e. they either evaluate to NA or NaN).

...
Arguments to be passed through to xtable and print.xtable.

env
The environment in which to evaluate expr.

quoted
Is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable.

outputArgs
A list of arguments to be passed through to the implicit call to tableOutput when renderTable is used in an interactive R Markdown document.

Details

The corresponding HTML output tag should be div and have the CSS class name shiny-html-output.

[Package shiny version 1.2.0 Index]