restoreInput {shiny}	R Documentation

Restore an input value

Description

This restores an input value from the current restore context. It should be called early on inside of input functions (like textInput).

Usage

restoreInput(id, default)

Arguments

id
Name of the input value to restore.

default
A default value to use, if there's no value to restore.

[Package shiny version 1.2.0 Index]