setBookmarkExclude {shiny}	R Documentation

Exclude inputs from bookmarking

Description

This function tells Shiny which inputs should be excluded from bookmarking. It should be called from inside the application's server function.

Usage

setBookmarkExclude(names = character(0), session = getDefaultReactiveDomain())

Arguments

names
A character vector containing names of inputs to exclude from bookmarking.

session
A shiny session object.

Details

This function can also be called from a module's server function, in which case it will exclude inputs with the specified names, from that module. It will not affect inputs from other modules or from the top level of the Shiny application.

See Also

enableBookmarking() for examples.

[Package shiny version 1.9.1 Index]