setSerializer {shiny}	R Documentation

Add a function for serializing an input before bookmarking application state

Description

Add a function for serializing an input before bookmarking application state

Usage

setSerializer(inputId, fun, session = getDefaultReactiveDomain())

Arguments

inputId
Name of the input value.

fun
A function that takes the input value and returns a modified value. The returned value will be used for the test snapshot.

session
A Shiny session object.

[Package shiny version 1.2.0 Index]