shiny-options {shiny}	R Documentation

Global options for Shiny

Description

There are a number of global options that affect Shiny's behavior. These can be set with (for example) options(shiny.trace=TRUE).

Details

shiny.launch.browser

A boolean which controls the default behavior when an app is run. See runApp for more information.

shiny.port

A port number that Shiny will listen on. See runApp for more information.

shiny.trace

Print messages sent between the R server and the web browser client to the R console. This is useful for debugging. Possible values are "send" (only print messages sent to the client), "recv" (only print messages received by the server), TRUE (print all messages), or FALSE (default; don't print any of these messages).

shiny.autoreload

If TRUE when a Shiny app is launched, the app directory will be continually monitored for changes to files that have the extensions: r, htm, html, js, css, png, jpg, jpeg, gif. If any changes are detected, all connected Shiny sessions are reloaded. This allows for fast feedback loops when tweaking Shiny UI.

Since monitoring for changes is expensive (we simply poll for last modified times), this feature is intended only for development.

You can customize the file patterns Shiny will monitor by setting the shiny.autoreload.pattern option. For example, to monitor only ui.R: options(shiny.autoreload.pattern = glob2rx("ui.R"))

The default polling interval is 500 milliseconds. You can change this by setting e.g. options(shiny.autoreload.interval = 2000) (every two seconds).

shiny.reactlog

If TRUE, enable logging of reactive events, which can be viewed later with the showReactLog function. This incurs a substantial performance penalty and should not be used in production.

shiny.usecairo

This is used to disable graphical rendering by the Cairo package, if it is installed. See plotPNG for more information.

shiny.maxRequestSize

This is a number which specifies the maximum web request size, which serves as a size limit for file uploads. If unset, the maximum request size defaults to 5MB.

shiny.suppressMissingContextError

Normally, invoking a reactive outside of a reactive context (or isolate()) results in an error. If this is TRUE, don't error in these cases. This should only be used for debugging or demonstrations of reactivity at the console.

shiny.host

The IP address that Shiny should listen on. See runApp for more information.

shiny.json.digits

The number of digits to use when converting numbers to JSON format to send to the client web browser.

shiny.minified

If this is TRUE or unset (the default), then Shiny will use minified JavaScript (shiny.min.js). If FALSE, then Shiny will use the un-minified JavaScript (shiny.js); this can be useful during development.

shiny.error

This can be a function which is called when an error occurs. For example, options(shiny.error=recover) will result a the debugger prompt when an error occurs.

shiny.table.class

CSS class names to use for tables.

shiny.deprecation.messages

This controls whether messages for deprecated functions in Shiny will be printed. See shinyDeprecated for more information.

shiny.fullstacktrace

Controls whether "pretty" or full stack traces are dumped to the console when errors occur during Shiny app execution. The default is FALSE (pretty stack traces).

shiny.stacktraceoffset

If TRUE, then Shiny's printed stack traces will display srcrefs one line above their usual location. This is an arguably more intuitive arrangement for casual R users, as the name of a function appears next to the srcref where it is defined, rather than where it is currently being called from.

shiny.sanitize.errors

If TRUE, then normal errors (i.e. errors not wrapped in safeError) won't show up in the app; a simple generic error message is printed instead (the error and strack trace printed to the console remain unchanged). The default is FALSE (unsanitized errors).If you want to sanitize errors in general, but you DO want a particular error e to get displayed to the user, then set this option to TRUE and use stop(safeError(e)) for errors you want the user to see.

shiny.testmode

If TRUE, then enable features for testing Shiny applications. If FALSE (the default), do not enable those features.

[Package shiny version 1.2.0 Index]