Web Application Framework for R

Shiny makes it incredibly easy to build interactive web applications with R. Automatic "reactive" binding between inputs and outputs and extensive prebuilt widgets make it possible to build beautiful, responsive, and powerful applications with minimal effort.

Details

The Shiny tutorial at https://shiny.rstudio.com/tutorial/ explains the framework in depth, walks you through building a simple application, and includes extensive annotated examples.

See also

Author

Maintainer: Winston Chang winston@posit.co (ORCID)

Authors:

Joe Cheng joe@posit.co

JJ Allaire jj@posit.co

Carson Sievert carson@posit.co (ORCID)

Barret Schloerke barret@posit.co (ORCID)

Yihui Xie yihui@posit.co

Jeff Allen

Jonathan McPherson jonathan@posit.co

Alan Dipert

Barbara Borges

Other contributors:

Posit Software, PBC [copyright holder, funder]

jQuery Foundation (jQuery library and jQuery UI library) [copyright holder]

jQuery contributors (jQuery library; authors listed in inst/www/shared/jquery-AUTHORS.txt) [contributor, copyright holder]

jQuery UI contributors (jQuery UI library; authors listed in inst/www/shared/jqueryui/AUTHORS.txt) [contributor, copyright holder]

Mark Otto (Bootstrap library) [contributor]

Jacob Thornton (Bootstrap library) [contributor]

Bootstrap contributors (Bootstrap library) [contributor]

Twitter, Inc (Bootstrap library) [copyright holder]

Prem Nawaz Khan (Bootstrap accessibility plugin) [contributor]

Victor Tsaran (Bootstrap accessibility plugin) [contributor]

Dennis Lembree (Bootstrap accessibility plugin) [contributor]

Srinivasu Chakravarthula (Bootstrap accessibility plugin) [contributor]

Cathy O'Connor (Bootstrap accessibility plugin) [contributor]

PayPal, Inc (Bootstrap accessibility plugin) [copyright holder]

Stefan Petre (Bootstrap-datepicker library) [contributor, copyright holder]

Andrew Rowls (Bootstrap-datepicker library) [contributor, copyright holder]

Brian Reavis (selectize.js library) [contributor, copyright holder]

Salmen Bejaoui (selectize-plugin-a11y library) [contributor, copyright holder]

Denis Ineshin (ion.rangeSlider library) [contributor, copyright holder]

Sami Samhuri (Javascript strftime library) [contributor, copyright holder]

SpryMedia Limited (DataTables library) [contributor, copyright holder]

John Fraser (showdown.js library) [contributor, copyright holder]

John Gruber (showdown.js library) [contributor, copyright holder]

Ivan Sagalaev (highlight.js library) [contributor, copyright holder]

R Core Team (tar implementation from R) [contributor, copyright holder]