getShinyOption {shiny}	R Documentation

Get or set Shiny options

Description

getShinyOption retrieves the value of a Shiny option. shinyOptions sets the value of Shiny options; it can also be used to return a list of all currently-set Shiny options.

Usage

getShinyOption(name, default = NULL)

shinyOptions(...)

Arguments

name
Name of an option to get.

default
Value to be returned if the option is not currently set.

...
Options to set, with the form name = value.

Details

There is a global option set, which is available by default. When a Shiny application is run with runApp, that option set is duplicated and the new option set is available for getting or setting values. If options are set from global.R, app.R, ui.R, or server.R, or if they are set from inside the server function, then the options will be scoped to the application. When the application exits, the new option set is discarded and the global option set is restored.

Examples

## Not run: 
shinyOptions(myOption = 10)
getShinyOption("myOption")

## End(Not run)

[Package shiny version 1.2.0 Index]