showModal {shiny}	R Documentation

Show or remove a modal dialog

Description

This causes a modal dialog to be displayed in the client browser, and is typically used with modalDialog().

Usage

showModal(ui, session = getDefaultReactiveDomain())

removeModal(session = getDefaultReactiveDomain())

Arguments

ui
UI content to show in the modal.

session
The session object passed to function given to shinyServer.

See Also

modalDialog() for examples.

[Package shiny version 1.9.1 Index]