sidebarPanel {shiny}	R Documentation

Create a sidebar panel

Description

Create a sidebar panel containing input controls that can in turn be passed to sidebarLayout.

Usage

sidebarPanel(..., width = 4)

Arguments

...
UI elements to include on the sidebar

width
The width of the sidebar. For fluid layouts this is out of 12 total units; for fixed layouts it is out of whatever the width of the sidebar's parent column is.

Value

A sidebar that can be passed to sidebarLayout

Examples

# Sidebar with controls to select a dataset and specify
# the number of observations to view
sidebarPanel(
  selectInput("dataset", "Choose a dataset:",
              choices = c("rock", "pressure", "cars")),

  numericInput("obs", "Observations:", 10)
)

[Package shiny version 1.2.0 Index]