snapshotPreprocessInput {shiny}	R Documentation

Add a function for preprocessing an input before taking a test snapshot

Description

Add a function for preprocessing an input before taking a test snapshot

Usage

snapshotPreprocessInput(inputId, fun, session = getDefaultReactiveDomain())

Arguments

inputId
Name of the input value.

fun
A function that takes the input value and returns a modified value. The returned value will be used for the test snapshot.

session
A Shiny session object.

[Package shiny version 1.9.1 Index]