tabPanel {shiny}	R Documentation

Create a tab panel

Description

Create a tab panel that can be included within a tabsetPanel.

Usage

tabPanel(title, ..., value = title, icon = NULL)

Arguments

title
Display title for tab

...
UI elements to include within the tab

value
The value that should be sent when tabsetPanel reports that this tab is selected. If omitted and tabsetPanel has an id, then the title will be used..

icon
Optional icon to appear on the tab. This attribute is only valid when using a tabPanel within a navbarPage.

Value

A tab that can be passed to tabsetPanel

See Also

tabsetPanel

Examples

# Show a tabset that includes a plot, summary, and
# table view of the generated distribution
mainPanel(
  tabsetPanel(
    tabPanel("Plot", plotOutput("plot")),
    tabPanel("Summary", verbatimTextOutput("summary")),
    tabPanel("Table", tableOutput("table"))
  )
)

[Package shiny version 1.2.0 Index]