Create a text input control

Create an input control for entry of unstructured text values

textInput(
  inputId,
  label,
  value = "",
  width = NULL,
  placeholder = NULL,
  ...,
  updateOn = c("change", "blur")
)

Arguments

inputId

The input slot that will be used to access the value.

label

Display label for the control, or NULL for no label.

value

Initial value.

width

The width of the input, e.g. '400px', or '100%'; see validateCssUnit().

placeholder

A character string giving the user a hint as to what can be entered into the control. Internet Explorer 8 and 9 do not support this option.

...

Ignored, included to require named arguments and for future feature expansion.

updateOn

A character vector specifying when the input should be updated. Options are "change" (default) and "blur". Use "change" to update the input immediately whenever the value changes. Use "blur"to delay the input update until the input loses focus (the user moves away from the input), or when Enter is pressed (or Cmd/Ctrl + Enter for textAreaInput()).

Value

A text input control that can be added to a UI definition.

Server value

A character string of the text input. The default value is "" unless value is provided.

See also

Examples

## Only run examples in interactive R sessions
if (interactive()) {

ui <- fluidPage(
  textInput("caption", "Caption", "Data Summary"),
  verbatimTextOutput("value")
)
server <- function(input, output) {
  output$value <- renderText({ input$caption })
}
shinyApp(ui, server)
}