updateDateRangeInput {shiny}	R Documentation

Change the start and end values of a date range input on the client

Description

Change the start and end values of a date range input on the client

Usage

updateDateRangeInput(session, inputId, label = NULL, start = NULL,
  end = NULL, min = NULL, max = NULL)

Arguments

session
The session object passed to function given to shinyServer.

inputId
The id of the input object.

label
The label to set for the input object.

start
The start date. Either a Date object, or a string in yyyy-mm-dd format. Supplying NA clears the start date.

end
The end date. Either a Date object, or a string in yyyy-mm-dd format. Supplying NA clears the end date.

min
The minimum allowed date. Either a Date object, or a string in yyyy-mm-dd format.

max
The maximum allowed date. Either a Date object, or a string in yyyy-mm-dd format.

Details

The input updater functions send a message to the client, telling it to change the settings of an input object. The messages are collected and sent after all the observers (including outputs) have finished running.

The syntax of these functions is similar to the functions that created the inputs in the first place. For example, numericInput() and updateNumericInput() take a similar set of arguments.

Any arguments with NULL values will be ignored; they will not result in any changes to the input object on the client.

For radioButtons(), checkboxGroupInput() and selectInput(), the set of choices can be cleared by using choices=character(0). Similarly, for these inputs, the selected item can be cleared by using selected=character(0).

See Also

dateRangeInput

Examples

## Only run examples in interactive R sessions
if (interactive()) {

ui <- fluidPage(
  sliderInput("n", "Day of month", 1, 30, 10),
  dateRangeInput("inDateRange", "Input date range")
)

server <- function(input, output, session) {
  observe({
    date <- as.Date(paste0("2013-04-", input$n))

    updateDateRangeInput(session, "inDateRange",
      label = paste("Date range label", input$n),
      start = date - 1,
      end = date + 1,
      min = date - 5,
      max = date + 5
    )
  })
}

shinyApp(ui, server)
}

[Package shiny version 1.2.0 Index]