verticalLayout {shiny}	R Documentation

Lay out UI elements vertically

Description

Create a container that includes one or more rows of content (each element passed to the container will appear on it's own line in the UI)

Usage

verticalLayout(..., fluid = TRUE)

Arguments

...
Elements to include within the container

fluid
TRUE to use fluid layout; FALSE to use fixed layout.

See Also

Other layout functions: fillPage(), fixedPage(), flowLayout(), fluidPage(), navbarPage(), sidebarLayout(), splitLayout()

Examples

## Only run examples in interactive R sessions
if (interactive()) {

ui <- fluidPage(
  verticalLayout(
    a(href="http://example.com/link1", "Link One"),
    a(href="http://example.com/link2", "Link Two"),
    a(href="http://example.com/link3", "Link Three")
  )
)
shinyApp(ui, server = function(input, output) { })
}

[Package shiny version 1.9.1 Index]