viewer {shiny}	R Documentation

Viewer options

Description

Use these functions to control where the gadget is displayed in RStudio (or other R environments that emulate RStudio's viewer pane/dialog APIs). If viewer APIs are not available in the current R environment, then the gadget will be displayed in the system's default web browser (see browseURL).

Usage

paneViewer(minHeight = NULL)

dialogViewer(dialogName, width = 600, height = 600)

browserViewer(browser = getOption("browser"))

Arguments

minHeight
The minimum height (in pixels) desired to show the gadget in the viewer pane. If a positive number, resize the pane if necessary to show at least that many pixels. If NULL, use the existing viewer pane size. If "maximize", use the maximum available vertical space.

dialogName
The window title to display for the dialog.

width, height
The desired dialog width/height, in pixels.

browser
See browseURL.

Value

A function that takes a single url parameter, suitable for passing as the viewer argument of runGadget.

[Package shiny version 1.2.0 Index]