withMathJax {shiny}	R Documentation

Load the MathJax library and typeset math expressions

Description

This function adds MathJax to the page and typeset the math expressions (if found) in the content .... It only needs to be called once in an app unless the content is rendered after the page is loaded, e.g. via renderUI(), in which case we have to call it explicitly every time we write math expressions to the output.

Usage

withMathJax(...)

Arguments

...
any HTML elements to apply MathJax to

Examples

withMathJax(helpText("Some math here $$\\alpha+\\beta$$"))
# now we can just write "static" content without withMathJax()
div("more math here $$\\sqrt{2}$$")

[Package shiny version 1.9.1 Index]