bookmarkButton {shiny}	R Documentation

Create a button for bookmarking/sharing

Description

A bookmarkButton is a actionButton with a default label that consists of a link icon and the text "Bookmark...". It is meant to be used for bookmarking state.

Usage

bookmarkButton(label = "Bookmark...", icon = shiny::icon("link", lib =
  "glyphicon"),
  title = "Bookmark this application's state and get a URL for sharing.",
  ..., id = "._bookmark_")

Arguments

label
The contents of the button or link–usually a text label, but you could also use any other HTML, like an image.

icon
An optional icon to appear on the button.

title
A tooltip that is shown when the mouse cursor hovers over the button.

...
Named attributes to be applied to the button or link.

id
An ID for the bookmark button. The only time it is necessary to set the ID unless you have more than one bookmark button in your application. If you specify an input ID, it should be excluded from bookmarking with setBookmarkExclude, and you must create an observer that does the bookmarking when the button is pressed. See the examples below.

See Also

enableBookmarking for more examples.

Examples

## Only run these examples in interactive sessions
if (interactive()) {

# This example shows how to use multiple bookmark buttons. If you only need
# a single bookmark button, see examples in ?enableBookmarking.
ui <- function(request) {
  fluidPage(
    tabsetPanel(id = "tabs",
      tabPanel("One",
        checkboxInput("chk1", "Checkbox 1"),
        bookmarkButton(id = "bookmark1")
      ),
      tabPanel("Two",
        checkboxInput("chk2", "Checkbox 2"),
        bookmarkButton(id = "bookmark2")
      )
    )
  )
}
server <- function(input, output, session) {
  # Need to exclude the buttons from themselves being bookmarked
  setBookmarkExclude(c("bookmark1", "bookmark2"))

  # Trigger bookmarking with either button
  observeEvent(input$bookmark1, {
    session$doBookmark()
  })
  observeEvent(input$bookmark2, {
    session$doBookmark()
  })
}
enableBookmarking(store = "url")
shinyApp(ui, server)
}

[Package shiny version 1.2.0 Index]