bootstrapLib {shiny}	R Documentation

Bootstrap libraries

Description

This function returns a set of web dependencies necessary for using Bootstrap components in a web page.

Usage

bootstrapLib(theme = NULL)

Arguments

theme
Alternative Bootstrap stylesheet (normally a css file within the www directory, e.g. www/bootstrap.css)

Details

It isn't necessary to call this function if you use bootstrapPage or others which use bootstrapPage, such basicPage, fluidPage, fillPage, pageWithSidebar, and navbarPage, because they already include the Bootstrap web dependencies.

[Package shiny version 1.2.0 Index]