bootstrapPage {shiny}	R Documentation

Create a Bootstrap page

Description

Create a Shiny UI page that loads the CSS and JavaScript for Bootstrap, and has no content in the page body (other than what you provide).

Usage

bootstrapPage(..., title = NULL, responsive = NULL, theme = NULL)

basicPage(...)

Arguments

...
The contents of the document body.

title
The browser window title (defaults to the host URL of the page)

responsive
This option is deprecated; it is no longer optional with Bootstrap 3.

theme
Alternative Bootstrap stylesheet (normally a css file within the www directory, e.g. www/bootstrap.css)

Details

This function is primarily intended for users who are proficient in HTML/CSS, and know how to lay out pages in Bootstrap. Most applications should use fluidPage along with layout functions like fluidRow and sidebarLayout.

Value

A UI defintion that can be passed to the shinyUI function.

Note

The basicPage function is deprecated, you should use the fluidPage function instead.

See Also

fluidPage, fixedPage

[Package shiny version 1.2.0 Index]