callModule {shiny}	R Documentation

Invoke a Shiny module

Description

Shiny's module feature lets you break complicated UI and server logic into smaller, self-contained pieces. Compared to large monolithic Shiny apps, modules are easier to reuse and easier to reason about. See the article at http://shiny.rstudio.com/articles/modules.html to learn more.

Usage

callModule(module, id, ..., session = getDefaultReactiveDomain())

Arguments

module
A Shiny module server function

id
An ID string that corresponds with the ID used to call the module's UI function

...
Additional parameters to pass to module server function

session
Session from which to make a child scope (the default should almost always be used)

Value

The return value, if any, from executing the module server function

See Also

http://shiny.rstudio.com/articles/modules.html

[Package shiny version 1.2.0 Index]