callModule {shiny}	R Documentation

Invoke a Shiny module

Description

Note: As of Shiny 1.5.0, we recommend using moduleServer() instead of callModule(), because the syntax is a little easier to understand, and modules created with moduleServer can be tested with testServer().

Usage

callModule(module, id, ..., session = getDefaultReactiveDomain())

Arguments

module
A Shiny module server function

id
An ID string that corresponds with the ID used to call the module's UI function

...
Additional parameters to pass to module server function

session
Session from which to make a child scope (the default should almost always be used)

Value

The return value, if any, from executing the module server function

[Package shiny version 1.6.0 Index]