clickOpts {shiny}	R Documentation

Control interactive plot point events

Description

These functions give control over the click, dblClick and hover events generated by imageOutput() and plotOutput().

Usage

clickOpts(id, clip = TRUE)

dblclickOpts(id, clip = TRUE, delay = 400)

hoverOpts(
  id,
  delay = 300,
  delayType = c("debounce", "throttle"),
  clip = TRUE,
  nullOutside = TRUE
)

Arguments

id
Input value name. For example, if the value is "plot_click", then the event data will be available as input$plot_click.

clip
Should the click area be clipped to the plotting area? If FALSE, then the server will receive click events even when the mouse is outside the plotting area, as long as it is still inside the image.

delay
For dblClickOpts(): the maximum delay (in ms) between a pair clicks for them to be counted as a double-click.

For hoverOpts(): how long to delay (in ms) when debouncing or throttling before sending the mouse location to the server.

delayType
The type of algorithm for limiting the number of hover events. Use "throttle" to limit the number of hover events to one every delay milliseconds. Use "debounce" to suspend events while the cursor is moving, and wait until the cursor has been at rest for delay milliseconds before sending an event.

nullOutside
If TRUE (the default), the value will be set to NULL when the mouse exits the plotting area. If FALSE, the value will stop changing when the cursor exits the plotting area.

See Also

brushOpts() for brushing events.

[Package shiny version 1.6.0 Index]