getCurrentTheme {shiny}	R Documentation

Obtain Shiny's Bootstrap Sass theme

Description

Intended for use by Shiny developers to create Shiny bindings with intelligent styling based on the bootstrapLib()'s theme value.

Usage

getCurrentTheme()

Value

If called at render-time (i.e., inside a htmltools::tagFunction()), and bootstrapLib()'s theme has been set to a bslib::bs_theme() object, then this returns the theme. Otherwise, this returns NULL.

See Also

getCurrentOutputInfo(), bootstrapLib(), htmltools::tagFunction()

[Package shiny version 1.6.0 Index]