htmlOutput {shiny}	R Documentation

Create an HTML output element

Description

Render a reactive output variable as HTML within an application page. The text will be included within an HTML div tag, and is presumed to contain HTML content which should not be escaped.

Usage

htmlOutput(outputId, inline = FALSE, container = if (inline) span else
  div, ...)

uiOutput(outputId, inline = FALSE, container = if (inline) span else
  div, ...)

Arguments

outputId
output variable to read the value from

inline
use an inline (span()) or block container (div()) for the output

container
a function to generate an HTML element to contain the text

...
Other arguments to pass to the container tag function. This is useful for providing additional classes for the tag.

Details

uiOutput is intended to be used with renderUI on the server side. It is currently just an alias for htmlOutput.

Value

An HTML output element that can be included in a panel

Examples

htmlOutput("summary")

# Using a custom container and class
tags$ul(
  htmlOutput("summary", container = tags$li, class = "custom-li-output")
)

[Package shiny version 1.2.0 Index]