icon {shiny}	R Documentation

Create an icon

Description

Create an icon for use within a page. Icons can appear on their own, inside of a button, or as an icon for a tabPanel within a navbarPage.

Usage

icon(name, class = NULL, lib = "font-awesome")

Arguments

name
Name of icon. Icons are drawn from the Font Awesome Free (currently icons from the v5.3.1 set are supported with the v4 naming convention) and Glyphicons libraries. Note that the "fa-" and "glyphicon-" prefixes should not be used in icon names (i.e. the "fa-calendar" icon should be referred to as "calendar")

class
Additional classes to customize the style of the icon (see the usage examples for details on supported styles).

lib
Icon library to use ("font-awesome" or "glyphicon")

Value

An icon element

See Also

For lists of available icons, see http://fontawesome.io/icons/ and http://getbootstrap.com/components/#glyphicons.

Examples

# add an icon to a submit button
submitButton("Update View", icon = icon("refresh"))

navbarPage("App Title",
  tabPanel("Plot", icon = icon("bar-chart-o")),
  tabPanel("Summary", icon = icon("list-alt")),
  tabPanel("Table", icon = icon("table"))
)

[Package shiny version 1.2.0 Index]