loadSupport {shiny}	R Documentation

Load an app's supporting R files

Description

Loads all of the supporting R files of a Shiny application. Specifically, this function loads any top-level supporting .R files in the R/ directory adjacent to the app.R/server.R/ui.R files.

Usage

loadSupport(
  appDir = NULL,
  renv = new.env(parent = globalenv()),
  globalrenv = globalenv()
)

Arguments

appDir
The application directory. If appDir is NULL or not supplied, the nearest enclosing directory that is a Shiny app, starting with the current directory, is used.

renv
The environmeny in which the files in the R/ directory should be evaluated.

globalrenv
The environment in which global.R should be evaluated. If NULL, global.R will not be evaluated at all.

Details

Since Shiny 1.5.0, this function is called by default when running an application. If it causes problems, there are two ways to opt out. You can either place a file named _disable_autoload.R in your R/ directory, or set options(shiny.autoload.r=FALSE). If you set this option, it will affect any application that runs later in the same R session, potentially breaking it, so after running your application, you should unset option with options(shiny.autoload.r=NULL)

The files are sourced in alphabetical order (as determined by list.files). global.R is evaluated before the supporting R files in the R/ directory.

[Package shiny version 1.6.0 Index]