makeReactiveBinding {shiny}	R Documentation

Make a reactive variable

Description

Turns a normal variable into a reactive variable, that is, one that has reactive semantics when assigned or read in the usual ways. The variable may already exist; if so, its value will be used as the initial value of the reactive variable (or NULL if the variable did not exist).

Usage

makeReactiveBinding(symbol, env = parent.frame())

Arguments

symbol
A character string indicating the name of the variable that should be made reactive

env
The environment that will contain the reactive variable

Value

None.

Examples

## Not run: 
a <- 10
makeReactiveBinding("a")
b <- reactive(a * -1)
observe(print(b()))
a <- 20

## End(Not run)

[Package shiny version 1.2.0 Index]