markOutputAttrs {shiny}	R Documentation

Mark a render function with attributes that will be used by the output

Description

Mark a render function with attributes that will be used by the output

Usage

markOutputAttrs(renderFunc, snapshotExclude = NULL,
  snapshotPreprocess = NULL)

Arguments

renderFunc
A function that is suitable for assigning to a Shiny output slot.

snapshotExclude
If TRUE, exclude the output from test snapshots.

snapshotPreprocess
A function for preprocessing the value before taking a test snapshot.

[Package shiny version 1.2.0 Index]