Evaluate an expression without a reactive context

Temporarily blocks the current reactive context and evaluates the given expression. Any attempt to directly access reactive values or expressions in expr will give the same results as doing it at the top-level (by default, an error).

maskReactiveContext(expr)

Arguments

expr

An expression to evaluate.

Value

The value of expr.

See also