Create memory cache (deprecated)

Create memory cache (deprecated)

memoryCache(
  max_size = 200 * 1024^2,
  max_age = Inf,
  max_n = Inf,
  evict = c("lru", "fifo"),
  missing = key_missing(),
  exec_missing = deprecated(),
  logfile = NULL
)

Arguments

max_size

Maximum size of the cache, in bytes. If the cache exceeds this size, cached objects will be removed according to the value of the evict. Use Inf for no size limit. The default is 512 megabytes.

max_age

Maximum age of files in cache before they are evicted, in seconds. Use Inf for no age limit.

max_n

Maximum number of objects in the cache. If the number of objects exceeds this value, then cached objects will be removed according to the value of evict. Use Inf for no limit of number of items.

evict

The eviction policy to use to decide which objects are removed when a cache pruning occurs. Currently, "lru" and "fifo" are supported.

missing

A value to return when get(key) is called but the key is not present in the cache. The default is a key_missing() object. It is actually an expression that is evaluated each time there is a cache miss. See section Missing keys for more information.

exec_missing

Deprecated.

logfile

An optional filename or connection object to where logging information will be written. To log to the console, use stderr() or stdout().