navbarPage {shiny}	R Documentation

Create a page with a top level navigation bar

Description

Create a page that contains a top level navigation bar that can be used to toggle a set of tabPanel elements.

Usage

navbarPage(title, ..., id = NULL, selected = NULL,
  position = c("static-top", "fixed-top", "fixed-bottom"),
  header = NULL, footer = NULL, inverse = FALSE,
  collapsible = FALSE, collapsable, fluid = TRUE, responsive = NULL,
  theme = NULL, windowTitle = title)

navbarMenu(title, ..., menuName = title, icon = NULL)

Arguments

title
The title to display in the navbar

...
tabPanel elements to include in the page. The navbarMenu function also accepts strings, which will be used as menu section headers. If the string is a set of dashes like "----" a horizontal separator will be displayed in the menu.

id
If provided, you can use input$id in your server logic to determine which of the current tabs is active. The value will correspond to the value argument that is passed to tabPanel.

selected
The value (or, if none was supplied, the title) of the tab that should be selected by default. If NULL, the first tab will be selected.

position
Determines whether the navbar should be displayed at the top of the page with normal scrolling behavior ("static-top"), pinned at the top ("fixed-top"), or pinned at the bottom ("fixed-bottom"). Note that using "fixed-top" or "fixed-bottom" will cause the navbar to overlay your body content, unless you add padding, e.g.: tags$style(type="text/css", "body {padding-top: 70px;}")

header
Tag or list of tags to display as a common header above all tabPanels.

footer
Tag or list of tags to display as a common footer below all tabPanels

inverse
TRUE to use a dark background and light text for the navigation bar

collapsible
TRUE to automatically collapse the navigation elements into a menu when the width of the browser is less than 940 pixels (useful for viewing on smaller touchscreen device)

collapsable
Deprecated; use collapsible instead.

fluid
TRUE to use a fluid layout. FALSE to use a fixed layout.

responsive
This option is deprecated; it is no longer optional with Bootstrap 3.

theme
Alternative Bootstrap stylesheet (normally a css file within the www directory). For example, to use the theme located at www/bootstrap.css you would use theme = "bootstrap.css".

windowTitle
The title that should be displayed by the browser window. Useful if title is not a string.

menuName
A name that identifies this navbarMenu. This is needed if you want to insert/remove or show/hide an entire navbarMenu.

icon
Optional icon to appear on a navbarMenu tab.

Details

The navbarMenu function can be used to create an embedded menu within the navbar that in turns includes additional tabPanels (see example below).

Value

A UI defintion that can be passed to the shinyUI function.

See Also

tabPanel, tabsetPanel, updateNavbarPage, insertTab, showTab

Examples

navbarPage("App Title",
  tabPanel("Plot"),
  tabPanel("Summary"),
  tabPanel("Table")
)

navbarPage("App Title",
  tabPanel("Plot"),
  navbarMenu("More",
    tabPanel("Summary"),
    "----",
    "Section header",
    tabPanel("Table")
  )
)

[Package shiny version 1.2.0 Index]