navlistPanel {shiny}	R Documentation

Create a navigation list panel

Description

Create a navigation list panel that provides a list of links on the left which navigate to a set of tabPanels displayed to the right.

Usage

navlistPanel(
  ...,
  id = NULL,
  selected = NULL,
  header = NULL,
  footer = NULL,
  well = TRUE,
  fluid = TRUE,
  widths = c(4, 8)
)

Arguments

...
tabPanel() elements to include in the navlist

id
If provided, you can use input$id in your server logic to determine which of the current navlist items is active. The value will correspond to the value argument that is passed to tabPanel().

selected
The value (or, if none was supplied, the title) of the navigation item that should be selected by default. If NULL, the first navigation will be selected.

header
Tag or list of tags to display as a common header above all tabPanels.

footer
Tag or list of tags to display as a common footer below all tabPanels

well
TRUE to place a well (gray rounded rectangle) around the navigation list.

fluid
TRUE to use fluid layout; FALSE to use fixed layout.

widths
Column widths of the navigation list and tabset content areas respectively.

Details

You can include headers within the navlistPanel by including plain text elements in the list. Versions of Shiny before 0.11 supported separators with "——", but as of 0.11, separators were no longer supported. This is because version 0.11 switched to Bootstrap 3, which doesn't support separators.

See Also

tabPanel(), updateNavlistPanel(), insertTab(), showTab()

Examples

fluidPage(

  titlePanel("Application Title"),

  navlistPanel(
    "Header",
    tabPanel("First"),
    tabPanel("Second"),
    tabPanel("Third")
  )
)

<!–html_preserve–>

Application Title

Header

First

Second

Third

<!–/html_preserve–>

[Package shiny version 1.7.1 Index]