numericInput {shiny}	R Documentation

Create a numeric input control

Description

Create an input control for entry of numeric values

Usage

numericInput(inputId, label, value, min = NA, max = NA, step = NA,
  width = NULL)

Arguments

inputId
The input slot that will be used to access the value.

label
Display label for the control, or NULL for no label.

value
Initial value.

min
Minimum allowed value

max
Maximum allowed value

step
Interval to use when stepping between min and max

width
The width of the input, e.g. '400px', or '100%'; see validateCssUnit.

Value

A numeric input control that can be added to a UI definition.

See Also

updateNumericInput

Other input elements: actionButton, checkboxGroupInput, checkboxInput, dateInput, dateRangeInput, fileInput, passwordInput, radioButtons, selectInput, sliderInput, submitButton, textAreaInput, textInput, varSelectInput

Examples

## Only run examples in interactive R sessions
if (interactive()) {

ui <- fluidPage(
  numericInput("obs", "Observations:", 10, min = 1, max = 100),
  verbatimTextOutput("value")
)
server <- function(input, output) {
  output$value <- renderText({ input$obs })
}
shinyApp(ui, server)
}

[Package shiny version 1.2.0 Index]