onFlush {shiny}	R Documentation

Add callbacks for Shiny session events

Description

These functions are for registering callbacks on Shiny session events. onFlush registers a function that will be called before Shiny flushes the reactive system. onFlushed registers a function that will be called after Shiny flushes the reactive system. onSessionEnded registers a function to be called after the client has disconnected.

Usage

onFlush(fun, once = TRUE, session = getDefaultReactiveDomain())

onFlushed(fun, once = TRUE, session = getDefaultReactiveDomain())

onSessionEnded(fun, session = getDefaultReactiveDomain())

Arguments

fun
A callback function.

once
Should the function be run once, and then cleared, or should it re-run each time the event occurs. (Only for onFlush and onFlushed.)

session
A shiny session object.

Details

These functions should be called within the application's server function.

All of these functions return a function which can be called with no arguments to cancel the registration.

See Also

onStop() for registering callbacks that will be invoked when the application exits, or when a session ends.

[Package shiny version 1.2.0 Index]