outputOptions {shiny}	R Documentation

Set options for an output object.

Description

These are the available options for an output object:

suspendWhenHidden. When TRUE (the default), the output object will be suspended (not execute) when it is hidden on the web page. When FALSE, the output object will not suspend when hidden, and if it was already hidden and suspended, then it will resume immediately.

priority. The priority level of the output object. Queued outputs with higher priority values will execute before those with lower values.

Usage

outputOptions(x, name, ...)

Arguments

x
A shinyoutput object (typically output).

name
The name of an output observer in the shinyoutput object.

...
Options to set for the output observer.

Examples

## Not run: 
##D # Get the list of options for all observers within output
##D outputOptions(output)
##D 
##D # Disable suspend for output$myplot
##D outputOptions(output, "myplot", suspendWhenHidden = FALSE)
##D 
##D # Change priority for output$myplot
##D outputOptions(output, "myplot", priority = 10)
##D 
##D # Get the list of options for output$myplot
##D outputOptions(output, "myplot")
## End(Not run)

[Package shiny version 1.7.1 Index]