parseQueryString {shiny}	R Documentation

Parse a GET query string from a URL

Description

Returns a named list of key-value pairs.

Usage

parseQueryString(str, nested = FALSE)

Arguments

str
The query string. It can have a leading "?" or not.

nested
Whether to parse the query string of as a nested list when it contains pairs of square brackets []. For example, the query a[i1][j1]=x&b[i1][j1]=y&b[i2][j1]=z will be parsed as list(a = list(i1 = list(j1 = 'x')), b = list(i1 = list(j1 = 'y'), i2 = list(j1 = 'z'))) when nested = TRUE, and list(`a[i1][j1]` = 'x', `b[i1][j1]` = 'y', `b[i2][j1]` = 'z') when nested = FALSE.

Examples

parseQueryString("?foo=1&bar=b%20a%20r")

## $foo
## [1] "1"
## 
## $bar
## [1] "b a r"

## Not run: 
##D # Example of usage within a Shiny app
##D function(input, output, session) {
##D 
##D   output$queryText <- renderText({
##D     query <- parseQueryString(session$clientData$url_search)
##D 
##D     # Ways of accessing the values
##D     if (as.numeric(query$foo) == 1) {
##D       # Do something
##D     }
##D     if (query[["bar"]] == "targetstring") {
##D       # Do something else
##D     }
##D 
##D     # Return a string with key-value pairs
##D     paste(names(query), query, sep = "=", collapse=", ")
##D   })
##D }
## End(Not run)

[Package shiny version 1.7.1 Index]