Create an object for storing reactive values

This function returns an object for storing reactive values. It is similar to a list, but with special capabilities for reactive programming. When you read a value from it, the calling reactive expression takes a reactive dependency on that value, and when you write to it, it notifies any reactive functions that depend on that value. Note that values taken from the reactiveValues object are reactive, but the reactiveValues object itself is not.

reactiveValues(...)

Arguments

...

Objects that will be added to the reactivevalues object. All of these objects must be named.

See also

Examples

# Create the object with no values
values <- reactiveValues()

# Assign values to 'a' and 'b'
values$a <- 3
values[['b']] <- 4

if (FALSE) { # \dontrun{
# From within a reactive context, you can access values with:
values$a
values[['a']]
} # }

# If not in a reactive context (e.g., at the console), you can use isolate()
# to retrieve the value:
isolate(values$a)
#> [1] 3
isolate(values[['a']])
#> [1] 3

# Set values upon creation
values <- reactiveValues(a = 1, b = 2)
isolate(values$a)
#> [1] 1