Convert a reactivevalues object to a list

This function does something similar to what you might want or expect base::as.list() to do. The difference is that the calling context will take dependencies on every object in the reactivevalues object. To avoid taking dependencies on all the objects, you can wrap the call with isolate().

reactiveValuesToList(x, all.names = FALSE)

Arguments

x

A reactivevalues object.

all.names

If TRUE, include objects with a leading dot. If FALSE (the default) don't include those objects.

Examples

values <- reactiveValues(a = 1)
if (FALSE) { # \dontrun{
reactiveValuesToList(values)
} # }

# To get the objects without taking dependencies on them, use isolate().
# isolate() can also be used when calling from outside a reactive context (e.g.
# at the console)
isolate(reactiveValuesToList(values))
#> $a
#> [1] 1
#>