repeatable {shiny}	R Documentation

Make a random number generator repeatable

Description

Given a function that generates random data, returns a wrapped version of that function that always uses the same seed when called. The seed to use can be passed in explicitly if desired; otherwise, a random number is used.

Usage

repeatable(rngfunc, seed = stats::runif(1, 0, .Machine$integer.max))

Arguments

rngfunc
The function that is affected by the R session's seed.

seed
The seed to set every time the resulting function is called.

Value

A repeatable version of the function that was passed in.

Note

When called, the returned function attempts to preserve the R session's current seed by snapshotting and restoring base::.Random.seed().

Examples

rnormA <- repeatable(rnorm)
rnormB <- repeatable(rnorm)
rnormA(3)  # [1]  1.8285879 -0.7468041 -0.4639111

## [1]  0.7584342  0.7232627 -0.3269704

rnormA(3)  # [1]  1.8285879 -0.7468041 -0.4639111

## [1]  0.7584342  0.7232627 -0.3269704

rnormA(5)  # [1]  1.8285879 -0.7468041 -0.4639111 -1.6510126 -1.4686924

## [1]  0.7584342  0.7232627 -0.3269704 -0.7480580  0.2911230

rnormB(5)  # [1] -0.7946034  0.2568374 -0.6567597  1.2451387 -0.8375699

## [1] -1.6366811 -0.3065373  0.3180388  1.7366371 -0.3996917

[Package shiny version 1.7.1 Index]