addResourcePath {shiny}	R Documentation

Resource Publishing

Description

Add, remove, or list directory of static resources to Shiny's web server, with the given path prefix. Primarily intended for package authors to make supporting JavaScript/CSS files available to their components.

Usage

addResourcePath(prefix, directoryPath)

resourcePaths()

removeResourcePath(prefix)

Arguments

prefix
The URL prefix (without slashes). Valid characters are a-z, A-Z, 0-9, hyphen, period, and underscore. For example, a value of 'foo' means that any request paths that begin with '/foo' will be mapped to the given directory.

directoryPath
The directory that contains the static resources to be served.

Details

Shiny provides two ways of serving static files (i.e., resources):

Static files under the www/ directory are automatically made available under a request path that begins with /.

addResourcePath() makes static files in a directoryPath available under a request path that begins with prefix.

The second approach is primarily intended for package authors to make supporting JavaScript/CSS files available to their components.

Tools for managing static resources published by Shiny's web server:

addResourcePath() adds a directory of static resources.

resourcePaths() lists the currently active resource mappings.

removeResourcePath() removes a directory of static resources.

See Also

singleton()

Examples

addResourcePath('datasets', system.file('data', package='datasets'))
resourcePaths()

##                                                                                                                                                       shinyjs 
##       "/Users/jorge/Library/Caches/org.R-project.R/R/renv/cache/v5/R-4.1/aarch64-apple-darwin20/shinyjs/2.0.0/9ddfc91d4280eaa34c2103951538976f/shinyjs/srcjs" 
##                                                                                                                                              shiny-sass-1.7.1 
##                                                                  "/private/var/folders/j0/t1m1vgbd6jx466y6thpnfmzc0000gn/T/RtmpW6uUT1/shiny-sass6a52453e36dc" 
##                                                                                                                                               bootstrap-4.6.0 
##                                                                      "/private/var/folders/j0/t1m1vgbd6jx466y6thpnfmzc0000gn/T/RtmpW6uUT1/bslib-6a522fd4f18c" 
##                                                                                                                                               bs3compat-0.3.1 
##                                                                "/Users/jorge/Documents/R/Linaje/renv/library/R-4.1/aarch64-apple-darwin20/bslib/bs3compat/js" 
##                                                                                                                                   bootstrap-colorpicker-3.1.2 
##                                                          "/Users/jorge/Documents/R/Linaje/renv/library/R-4.1/aarch64-apple-darwin20/bslib/lib/bs-colorpicker" 
##                                                                                                                                               bs_themer-0.3.1 
##                                                                      "/Users/jorge/Documents/R/Linaje/renv/library/R-4.1/aarch64-apple-darwin20/bslib/themer" 
##                                                                                                                                           bs-themer-css-0.3.1 
##                                                               "/private/var/folders/j0/t1m1vgbd6jx466y6thpnfmzc0000gn/T/RtmpW6uUT1/bs-themer-css6a527783c907" 
##                                                                                                                                             htmlwidgets-1.5.4 
## "/Users/jorge/Library/Caches/org.R-project.R/R/renv/cache/v5/R-4.1/aarch64-apple-darwin20/htmlwidgets/1.5.4/76147821cd3fcd8c4b04e1ef0498e7fb/htmlwidgets/www" 
##                                                                                                                                                     vis-9.1.0 
##                                                    "/Users/jorge/Documents/R/Linaje/renv/library/R-4.1/aarch64-apple-darwin20/visNetwork/htmlwidgets/lib/vis" 
##                                                                                                                                      visNetwork-binding-2.1.0 
##                                                            "/Users/jorge/Documents/R/Linaje/renv/library/R-4.1/aarch64-apple-darwin20/visNetwork/htmlwidgets" 
##                                                                                                                                            font-awesome-4.7.0 
##                                      "/Users/jorge/Documents/R/Linaje/renv/library/R-4.1/aarch64-apple-darwin20/visNetwork/htmlwidgets/lib/fontawesome_4.7.0" 
##                                                                                                                                           font-awesome-5.15.3 
##                                                           "/Users/jorge/Documents/R/Linaje/renv/library/R-4.1/aarch64-apple-darwin20/fontawesome/fontawesome" 
##                                                                                                                                                      datasets 
##                                                                          "/Library/Frameworks/R.framework/Versions/4.1-arm64/Resources/library/datasets/data"

removeResourcePath('datasets')
resourcePaths()

##                                                                                                                                                       shinyjs 
##       "/Users/jorge/Library/Caches/org.R-project.R/R/renv/cache/v5/R-4.1/aarch64-apple-darwin20/shinyjs/2.0.0/9ddfc91d4280eaa34c2103951538976f/shinyjs/srcjs" 
##                                                                                                                                              shiny-sass-1.7.1 
##                                                                  "/private/var/folders/j0/t1m1vgbd6jx466y6thpnfmzc0000gn/T/RtmpW6uUT1/shiny-sass6a52453e36dc" 
##                                                                                                                                               bootstrap-4.6.0 
##                                                                      "/private/var/folders/j0/t1m1vgbd6jx466y6thpnfmzc0000gn/T/RtmpW6uUT1/bslib-6a522fd4f18c" 
##                                                                                                                                               bs3compat-0.3.1 
##                                                                "/Users/jorge/Documents/R/Linaje/renv/library/R-4.1/aarch64-apple-darwin20/bslib/bs3compat/js" 
##                                                                                                                                   bootstrap-colorpicker-3.1.2 
##                                                          "/Users/jorge/Documents/R/Linaje/renv/library/R-4.1/aarch64-apple-darwin20/bslib/lib/bs-colorpicker" 
##                                                                                                                                               bs_themer-0.3.1 
##                                                                      "/Users/jorge/Documents/R/Linaje/renv/library/R-4.1/aarch64-apple-darwin20/bslib/themer" 
##                                                                                                                                           bs-themer-css-0.3.1 
##                                                               "/private/var/folders/j0/t1m1vgbd6jx466y6thpnfmzc0000gn/T/RtmpW6uUT1/bs-themer-css6a527783c907" 
##                                                                                                                                             htmlwidgets-1.5.4 
## "/Users/jorge/Library/Caches/org.R-project.R/R/renv/cache/v5/R-4.1/aarch64-apple-darwin20/htmlwidgets/1.5.4/76147821cd3fcd8c4b04e1ef0498e7fb/htmlwidgets/www" 
##                                                                                                                                                     vis-9.1.0 
##                                                    "/Users/jorge/Documents/R/Linaje/renv/library/R-4.1/aarch64-apple-darwin20/visNetwork/htmlwidgets/lib/vis" 
##                                                                                                                                      visNetwork-binding-2.1.0 
##                                                            "/Users/jorge/Documents/R/Linaje/renv/library/R-4.1/aarch64-apple-darwin20/visNetwork/htmlwidgets" 
##                                                                                                                                            font-awesome-4.7.0 
##                                      "/Users/jorge/Documents/R/Linaje/renv/library/R-4.1/aarch64-apple-darwin20/visNetwork/htmlwidgets/lib/fontawesome_4.7.0" 
##                                                                                                                                           font-awesome-5.15.3 
##                                                           "/Users/jorge/Documents/R/Linaje/renv/library/R-4.1/aarch64-apple-darwin20/fontawesome/fontawesome"

# make sure all resources are removed
lapply(names(resourcePaths()), removeResourcePath)

## [[1]]
## [1] TRUE
## 
## [[2]]
## [1] TRUE
## 
## [[3]]
## [1] TRUE
## 
## [[4]]
## [1] TRUE
## 
## [[5]]
## [1] TRUE
## 
## [[6]]
## [1] TRUE
## 
## [[7]]
## [1] TRUE
## 
## [[8]]
## [1] TRUE
## 
## [[9]]
## [1] TRUE
## 
## [[10]]
## [1] TRUE
## 
## [[11]]
## [1] TRUE
## 
## [[12]]
## [1] TRUE

[Package shiny version 1.7.1 Index]