runExample {shiny}	R Documentation

Run Shiny Example Applications

Description

Launch Shiny example applications, and optionally, your system's web browser.

Usage

runExample(example = NA, port = NULL,
  launch.browser = getOption("shiny.launch.browser", interactive()),
  host = getOption("shiny.host", "127.0.0.1"), display.mode = c("auto",
  "normal", "showcase"))

Arguments

example
The name of the example to run, or NA (the default) to list the available examples.

port
The TCP port that the application should listen on. Defaults to choosing a random port.

launch.browser
If true, the system's default web browser will be launched automatically after the app is started. Defaults to true in interactive sessions only.

host
The IPv4 address that the application should listen on. Defaults to the shiny.host option, if set, or "127.0.0.1" if not.

display.mode
The mode in which to display the example. Defaults to showcase, but may be set to normal to see the example without code or commentary.

Examples

## Only run this example in interactive R sessions
if (interactive()) {
  # List all available examples
  runExample()

  # Run one of the examples
  runExample("01_hello")

  # Print the directory containing the code for all examples
  system.file("examples", package="shiny")
}

[Package shiny version 1.2.0 Index]