Run Shiny Example Applications

Launch Shiny example applications, and optionally, your system's web browser.

runExample(
  example = NA,
  port = getOption("shiny.port"),
  launch.browser = getOption("shiny.launch.browser", interactive()),
  host = getOption("shiny.host", "127.0.0.1"),
  display.mode = c("auto", "normal", "showcase"),
  package = "shiny"
)

Arguments

example

The name of the example to run, or NA (the default) to list the available examples.

port

The TCP port that the application should listen on. If the port is not specified, and the shiny.port option is set (with options(shiny.port = XX)), then that port will be used. Otherwise, use a random port between 3000:8000, excluding ports that are blocked by Google Chrome for being considered unsafe: 3659, 4045, 5060, 5061, 6000, 6566, 6665:6669 and 6697. Up to twenty random ports will be tried.

launch.browser

If true, the system's default web browser will be launched automatically after the app is started. Defaults to true in interactive sessions only.

host

The IPv4 address that the application should listen on. Defaults to the shiny.host option, if set, or "127.0.0.1" if not.

display.mode

The mode in which to display the example. Defaults to "auto", which uses the value of DisplayMode in the example's DESCRIPTION file. Set to "showcase" to show the app code and description with the running app, or "normal" to see the example without code or commentary.

package

The package in which to find the example (defaults to "shiny").

To provide examples in your package, store examples in the inst/examples-shiny directory of your package. Each example should be in its own subdirectory and should be runnable when runApp() is called on the subdirectory. Example apps can include a DESCRIPTION file and a README.md file to provide metadata and commentary about the example. See the article on Display Modes on the Shiny website for more information.

Examples

## Only run this example in interactive R sessions
if (interactive()) {
  # List all available examples
  runExample()

  # Run one of the examples
  runExample("01_hello")

  # Print the directory containing the code for all examples
  system.file("examples", package="shiny")
}