runGadget {shiny}	R Documentation

Run a gadget

Description

Similar to runApp, but handles input$cancel automatically, and if running in RStudio, defaults to viewing the app in the Viewer pane.

Usage

runGadget(app, server = NULL, port = getOption("shiny.port"),
  viewer = paneViewer(), stopOnCancel = TRUE)

Arguments

app
Either a Shiny app object as created by shinyApp et al, or, a UI object.

server
Ignored if app is a Shiny app object; otherwise, passed along to shinyApp (i.e. shinyApp(ui = app, server = server)).

port
See runApp.

viewer
Specify where the gadget should be displayed–viewer pane, dialog window, or external browser–by passing in a call to one of the viewer functions.

stopOnCancel
If TRUE (the default), then an observeEvent is automatically created that handles input$cancel by calling stopApp() with an error. Pass FALSE if you want to handle input$cancel yourself.

Value

The value returned by the gadget.

Examples

## Not run: 
library(shiny)

ui <- fillPage(...)

server <- function(input, output, session) {
  ...
}

# Either pass ui/server as separate arguments...
runGadget(ui, server)

# ...or as a single app object
runGadget(shinyApp(ui, server))

## End(Not run)

[Package shiny version 1.2.0 Index]