runTests {shiny}	R Documentation

Runs the tests associated with this Shiny app

Description

Sources the .R files in the top-level of tests/ much like R CMD check. These files are typically simple runners for tests nested in other directories under tests/.

Usage

runTests(appDir = ".", filter = NULL, assert = TRUE, envir = globalenv())

Arguments

appDir
The base directory for the application.

filter
If not NULL, only tests with file names matching this regular expression will be executed. Matching is performed on the file name including the extension.

assert
Logical value which determines if an error should be thrown if any error is captured.

envir
Parent testing environment in which to base the individual testing environments.

Details

Historically, shinytest recommended placing tests at the top-level of the tests/ directory. This older folder structure is not supported by runTests. Please see shinyAppTemplate() for more details.

Value

A data frame classed with the supplemental class "shiny_runtests". The data frame has the following columns:

Name 	 Type 	 Meaning

file 	 character(1) 	 File name of the runner script in tests/ that was sourced.

pass 	 logical(1) 	 Whether or not the runner script signaled an error when sourced.

result 	 any or NA 	 The return value of the runner

[Package shiny version 1.5.0 Index]