Print message for deprecated functions in Shiny

To disable these messages, use options(shiny.deprecation.messages=FALSE).

shinyDeprecated(
  version,
  what,
  with = NULL,
  details = NULL,
  type = c("deprecated", "superseded")
)

Arguments

version

Shiny version when the function was deprecated

what

Function with possible arguments

with

Possible function with arguments that should be used instead

details

Additional information to be added after a new line to the displayed message