shinyUI {shiny}	R Documentation

Create a Shiny UI handler

Description

Historically this function was used in ui.R files to register a user interface with Shiny. It is no longer required as of Shiny 0.10; simply ensure that the last expression to be returned from ui.R is a user interface. This function is kept for backwards compatibility with older applications. It returns the value that is passed to it.

Usage

shinyUI(ui)

Arguments

ui
A user interace definition

Value

The user interface definition, without modifications or side effects.

[Package shiny version 1.2.0 Index]