showBookmarkUrlModal {shiny}	R Documentation

Display a modal dialog for bookmarking

Description

This is a wrapper function for urlModal that is automatically called if an application is bookmarked but no other onBookmark callback was set. It displays a modal dialog with the bookmark URL, along with a subtitle that is appropriate for the type of bookmarking used ("url" or "server").

Usage

showBookmarkUrlModal(url)

Arguments

url
A URL to show in the modal dialog.

[Package shiny version 1.2.0 Index]