Create a sizing function that grows at a given ratio

Returns a function which takes a two-element vector representing an input width and height, and returns a two-element vector of width and height. The possible widths are the base width times the growthRate to any integer power. For example, with a base width of 500 and growth rate of 1.25, the possible widths include 320, 400, 500, 625, 782, and so on, both smaller and larger. Sizes are rounded up to the next pixel. Heights are computed the same way as widths.

sizeGrowthRatio(width = 400, height = 400, growthRate = 1.2)

Arguments

width, height

Base width and height.

growthRate

Growth rate multiplier.

See also

Examples

f <- sizeGrowthRatio(500, 500, 1.25)
f(c(400, 400))
#> [1] 500 500
f(c(500, 500))
#> [1] 500 500
f(c(530, 550))
#> [1] 625 625
f(c(625, 700))
#> [1] 625 782