sliderInput {shiny}	R Documentation

Slider Input Widget

Description

Constructs a slider widget to select a numeric value from a range.

Usage

sliderInput(inputId, label, min, max, value, step = NULL,
  round = FALSE, format = NULL, locale = NULL, ticks = TRUE,
  animate = FALSE, width = NULL, sep = ",", pre = NULL,
  post = NULL, timeFormat = NULL, timezone = NULL,
  dragRange = TRUE)

animationOptions(interval = 1000, loop = FALSE, playButton = NULL,
  pauseButton = NULL)

Arguments

inputId
The input slot that will be used to access the value.

label
Display label for the control, or NULL for no label.

min
The minimum value (inclusive) that can be selected.

max
The maximum value (inclusive) that can be selected.

value
The initial value of the slider. A numeric vector of length one will create a regular slider; a numeric vector of length two will create a double-ended range slider. A warning will be issued if the value doesn't fit between min and max.

step
Specifies the interval between each selectable value on the slider (if NULL, a heuristic is used to determine the step size). If the values are dates, step is in days; if the values are times (POSIXt), step is in seconds.

round
TRUE to round all values to the nearest integer; FALSE if no rounding is desired; or an integer to round to that number of digits (for example, 1 will round to the nearest 10, and -2 will round to the nearest .01). Any rounding will be applied after snapping to the nearest step.

format
Deprecated.

locale
Deprecated.

ticks
FALSE to hide tick marks, TRUE to show them according to some simple heuristics.

animate
TRUE to show simple animation controls with default settings; FALSE not to; or a custom settings list, such as those created using animationOptions.

width
The width of the input, e.g. '400px', or '100%'; see validateCssUnit.

sep
Separator between thousands places in numbers.

pre
A prefix string to put in front of the value.

post
A suffix string to put after the value.

timeFormat
Only used if the values are Date or POSIXt objects. A time format string, to be passed to the Javascript strftime library. See https://github.com/samsonjs/strftime for more details. The allowed format specifications are very similar, but not identical, to those for R's strftime function. For Dates, the default is "%F" (like "2015-07-01"), and for POSIXt, the default is "%F %T" (like "2015-07-01 15:32:10").

timezone
Only used if the values are POSIXt objects. A string specifying the time zone offset for the displayed times, in the format "+HHMM" or "-HHMM". If NULL (the default), times will be displayed in the browser's time zone. The value "+0000" will result in UTC time.

dragRange
This option is used only if it is a range slider (with two values). If TRUE (the default), the range can be dragged. In other words, the min and max can be dragged together. If FALSE, the range cannot be dragged.

interval
The interval, in milliseconds, between each animation step.

loop
TRUE to automatically restart the animation when it reaches the end.

playButton
Specifies the appearance of the play button. Valid values are a one-element character vector (for a simple text label), an HTML tag or list of tags (using tag and friends), or raw HTML (using HTML).

pauseButton
Similar to playButton, but for the pause button.

See Also

updateSliderInput

Other input elements: actionButton, checkboxGroupInput, checkboxInput, dateInput, dateRangeInput, fileInput, numericInput, passwordInput, radioButtons, selectInput, submitButton, textAreaInput, textInput, varSelectInput

Examples

## Only run examples in interactive R sessions
if (interactive()) {
options(device.ask.default = FALSE)

ui <- fluidPage(
  sliderInput("obs", "Number of observations:",
    min = 0, max = 1000, value = 500
  ),
  plotOutput("distPlot")
)

# Server logic
server <- function(input, output) {
  output$distPlot <- renderPlot({
    hist(rnorm(input$obs))
  })
}

# Complete app with UI and server components
shinyApp(ui, server)
}

[Package shiny version 1.2.0 Index]