snapshotPreprocessOutput {shiny}	R Documentation

Add a function for preprocessing an output before taking a test snapshot

Description

Add a function for preprocessing an output before taking a test snapshot

Usage

snapshotPreprocessOutput(x, fun)

Arguments

x
A reactive which will be assigned to an output.

fun
A function that takes the output value as an input and returns a modified value. The returned value will be used for the test snapshot.

[Package shiny version 1.2.0 Index]