tabsetPanel {shiny}	R Documentation

Create a tabset panel

Description

Create a tabset that contains tabPanel elements. Tabsets are useful for dividing output into multiple independently viewable sections.

Usage

tabsetPanel(..., id = NULL, selected = NULL, type = c("tabs",
  "pills"), position = NULL)

Arguments

...
tabPanel elements to include in the tabset

id
If provided, you can use input$id in your server logic to determine which of the current tabs is active. The value will correspond to the value argument that is passed to tabPanel.

selected
The value (or, if none was supplied, the title) of the tab that should be selected by default. If NULL, the first tab will be selected.

type
Use "tabs" for the standard look; Use "pills" for a more plain look where tabs are selected using a background fill color.

position
This argument is deprecated; it has been discontinued in Bootstrap 3.

Value

A tabset that can be passed to mainPanel

See Also

tabPanel, updateTabsetPanel, insertTab, showTab

Examples

# Show a tabset that includes a plot, summary, and
# table view of the generated distribution
mainPanel(
  tabsetPanel(
    tabPanel("Plot", plotOutput("plot")),
    tabPanel("Summary", verbatimTextOutput("summary")),
    tabPanel("Table", tableOutput("table"))
  )
)

[Package shiny version 1.2.0 Index]