textOutput {shiny}	R Documentation

Create a text output element

Description

Render a reactive output variable as text within an application page. The text will be included within an HTML div tag by default.

Usage

textOutput(outputId, container = if (inline) span else div,
  inline = FALSE)

Arguments

outputId
output variable to read the value from

container
a function to generate an HTML element to contain the text

inline
use an inline (span()) or block container (div()) for the output

Details

Text is HTML-escaped prior to rendering. This element is often used to display renderText output variables.

Value

A text output element that can be included in a panel

Examples

h3(textOutput("caption"))

[Package shiny version 1.2.0 Index]