Create a text output element

Render a reactive output variable as text within an application page. textOutput() is usually paired with renderText() and puts regular text in <div> or <span>; verbatimTextOutput() is usually paired with renderPrint() and provides fixed-width text in a <pre>.

textOutput(outputId, container = if (inline) span else div, inline = FALSE)

verbatimTextOutput(outputId, placeholder = FALSE)

Arguments

outputId

output variable to read the value from

container

a function to generate an HTML element to contain the text

inline

use an inline (span()) or block container (div()) for the output

placeholder

if the output is empty or NULL, should an empty rectangle be displayed to serve as a placeholder? (does not affect behavior when the output is nonempty)

Value

An output element for use in UI.

Details

In both functions, text is HTML-escaped prior to rendering.

Examples

## Only run this example in interactive R sessions
if (interactive()) {
  shinyApp(
    ui = basicPage(
      textInput("txt", "Enter the text to display below:"),
      textOutput("text"),
      verbatimTextOutput("verb")
    ),
    server = function(input, output) {
      output$text <- renderText({ input$txt })
      output$verb <- renderText({ input$txt })
    }
  )
}