titlePanel {shiny}	R Documentation

Create a panel containing an application title.

Description

Create a panel containing an application title.

Usage

titlePanel(title, windowTitle = title)

Arguments

title
An application title to display

windowTitle
The title that should be displayed by the browser window.

Details

Calling this function has the side effect of including a title tag within the head. You can also specify a page title explicitly using the 'title' parameter of the top-level page function.

Examples

## Only run examples in interactive R sessions
if (interactive()) {

ui <- fluidPage(
  titlePanel("Hello Shiny!")
)
shinyApp(ui, server = function(input, output) { })
}

[Package shiny version 1.2.0 Index]