Temporarily set OpenTelemetry collection level

withOtelCollect() temporarily sets the OpenTelemetry collection level for the duration of evaluating expr. localOtelCollect() sets the collection level for the remainder of the current function scope.

These functions are useful for temporarily controlling telemetry collection during reactive expression creation. Only the following levels are allowed:

"none" - No telemetry data collected

"reactivity" - Collect reactive execution spans (includes session and reactive update events)

"all" - All available telemetry (currently equivalent to "reactivity")

Note that "session" and "reactive_update" levels are not permitted as these are runtime-specific levels that should only be set permanently via options(shiny.otel.collect = ...) or the SHINY_OTEL_COLLECT environment variable, not temporarily during reactive expression creation.

withOtelCollect(collect, expr)

localOtelCollect(collect, envir = parent.frame())

Arguments

collect

Character string specifying the OpenTelemetry collection level. Must be one of "none", "reactivity", or "all".

expr

Expression to evaluate with the specified collection level (for withOtelCollect()).

envir

Environment where the collection level should be set (for localOtelCollect()). Defaults to the parent frame.

Value

withOtelCollect() returns the value of expr.

localOtelCollect() is called for its side effect and returns the previous collect value invisibly.

Intended Usage

These functions are designed to perform sweeping changes to telemetry collection, such as enabling or disabling OpenTelemetry for an entire module or section of code where reactive expressions are being created:

# Enable telemetry for an entire module
withOtelCollect("all", {
  my_result <- my_module("my_id")
})

# Disable telemetry for expensive development-only reactives
withOtelCollect("none", {
  debug_reactive <- reactive({ expensive_debug_computation() })
})

Pipe Usage (Not Recommended)

While using withOtelCollect() as a pipe-able method, it is not recommended due to the use case where the reactive object is created before the withOtelCollect() call. In such cases, the reactive object will not inherit the intended OpenTelemetry settings.

Therefore, to avoid this hard-to-debug situation, we recommend that you only create your reactive objects within the withOtelCollect() call or after setting the local collection level with localOtelCollect().

# Technically works, but not recommended
x <- reactive({ ... }) %>% withOtelCollect(collect = "all")
x <- reactive({ ... }) |> withOtelCollect(collect = "all")
# Equivalent to:
x <- withOtelCollect("all", reactive({ ... }))

# Does NOT work as intended
x <- reactive({ ... })
# `x` was created outside of `withOtelCollect()`,
# therefore no OTel settings are applied
x_no_otel <- withOtelCollect("all", x)

# Best practice: Create the reactive object within `expr=`
withOtelCollect("all", {
  x_with_otel <- reactive({ ... })
  y_with_otel <- reactive({ ... })
})

See also

Examples

if (FALSE) { # \dontrun{
# Temporarily disable telemetry collection
withOtelCollect("none", {
  # Code here won't generate telemetry
  reactive({ input$x + 1 })
})

# Collect reactivity telemetry but not other events
withOtelCollect("reactivity", {
  # Reactive execution will be traced
  observe({ print(input$x) })
})

# Use local variant in a function
my_function <- function() {
  localOtelCollect("none")
  # Rest of function executes without telemetry
  reactive({ input$y * 2 })
}
} # }